Subject: AuraEffectActor
In the AuraEffectActor class, the OnOverlap function has been implemented.
This function is triggered when an overlap event occurs. It checks if the overlapping actor implements the IAbilitySystemInterface.
If it does, it retrieves the AuraAttributeSet from the actor's AbilitySystemComponent and increases the Health attribute by 25.
After applying the effect, the AuraEffectActor is destroyed.
The OnOverlap function is bound to the OnComponentBeginOverlap event in the BeginPlay function.

We used const_cast to cast the AuraAttributeSet to a non-const pointer. This is bad practice and should be avoided.
Typically we want to use a Gameplay Effect to modify attributes. That way the attribute set can remain const and we don't have to use const_cast.