Subject: Connected and Initialized the AuraHUD, AuraWidgetController base class, OverlayWidgetController, OverlayWidget.
We created in our HUD Class the Overlay Widget and Overlay Widget Controller base class.
In the Base Class we created a struct to store all of our primary classes we need transfer data from
model to the view. We created a function to initialize the Controller and Widget and called it from
the AuraCharacter class.

Review Code for more depth.

This was a very complex construction of code for setup, but follows proper coding standards in Unreal Engine.