Subject: Created Delegates to Update Health Globe and ManaGlobe on changes.
We created a delegate that changes our Health Globe Widget and our Mana Globe Widget when the attribute is changed.

In Blueprints we Cast from WidgetController to OverlayWidgetController.
We bind the Controller to Delegate it should listen to, and when the attribute changes,
should update the Health and Mana values.