Subject: Created a PreAttributeBaseChange and PostGameplayEffectExecute Function
These functions are used to make some changes before and after an attribute changes respectively.

In PostGameplayEffectExecute:
A Data Reference is passed in that allows us to access A LOT of data about the source and the target character. We can
get data like the abilitysystemcomponent of each, the characters, and the controllers of each from this data reference.

We created a struct to store a bunch of key data points:

```
USTRUCT()
struct FEffectProperties
{
	GENERATED_BODY()

	FEffectProperties(){} // Default Constructor

	FGameplayEffectContextHandle EffectContextHandle;

	/*Source*/
	UPROPERTY()
	UAbilitySystemComponent* SourceASC = nullptr;

	UPROPERTY()
	AActor* SourceAvatarActor = nullptr;

	UPROPERTY()
	AController* SourceController = nullptr;

	UPROPERTY()
	ACharacter* SourceCharacter = nullptr;

	/*Target*/
	UPROPERTY()
	UAbilitySystemComponent* TargetASC = nullptr;

	UPROPERTY()
	AActor* TargetAvatarActor = nullptr;

	UPROPERTY()
	AController* TargetController = nullptr;

	UPROPERTY()
	ACharacter* TargetCharacter = nullptr;

};

```
We then refactor our function to set the struct in PostGameplayEffectExecute.

This function can be used a lot in order to help adjust attributes after theyve been changed.