We are creating the system in which when attributes change on the ASC, when the ASC broadcasts, our AttributeMenuWidgetController class can view the changed attributes and their respective tags, then package up a DataAsset and broadcast is to our AttributeMenu widget.
To do that, an efficient way of handling tags is to create them in C++.

We are creating the tags in an AuraGameplayTags struct file. This struct is a Singleton.
We will have a function that initializes the tags and another function that allows us to fetch,
all of the tags.