Created the Data Asset that will store an array of GameplayTags using a Struct with each GameplayTags information.
We also started to create our own Blueprint Library.
We created a C++ class based on BlueprintLibrary.
The functions in these classes are typically Static and so they require a WorldContextObject.
We created an initial function that retrieves the Overlay WidgetController for a player and/or crafts one using the
    AuraHUD Class function GetOverlayWidgetController(Params)