Commit Message:
"Implemented delegate structure in AttributeMenuWidgetController

This commit introduces a delegate structure in the AttributeMenuWidgetController. The delegate, named
FAttributeInfoSignature, is declared using the DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam macro and is
designed to broadcast the initial value of an attribute to the UI.

The delegate is defined as a UPROPERTY, making it accessible from Blueprints and allowing functions to
be assigned to it from the Blueprint editor.

The BroadcastInitialValue function uses the delegate to broadcast the initial value of an attribute.
The FAuraAttributeInfo struct, which contains the initial value of the attribute, is passed as an argument
to the delegate's Broadcast function."

Blueprints:
In Blueprints we have to set out AttributeInfo to the DA_AttributeInfo
In AttributeMenu Graph in the Event Construct we need to Bind to the BroadcastInitialValue Node (This node muse be BlueprintCallable, set from the Widgets Parent C++ class.)
In the WidgetRowButton Graph we Bind the EventConstruct function to the AttributeInfoDelegate.
We create functions for setting both the Text of the WidgetRow and Value to be modifiable.
We Then Use the data from the AuraAttributeInfo to Set the Text and the Value.