Created an InputComponent class that will handle all of our inputs, instead of storing the bindings in our player controller class.
Created a Template function that can recieve function ptrs as inputs.
The functions we will pass in our functions for Pressed, Held, and Released Events.
We loop over our Array of InputActions and Bind each action to each one of these events.
That will give each action a Pressed, Held, and Released Event.