We call the BindAbilityActions() in our SetupInputComponent function to ensure all of our InputActions (Buttons for abilities) have all 3 Trigger Event function options.
We want that possiblity so that we can rebind Gameplay Abilities to different InputAction at runtime,
and so that we only need 1 function to manage InputActions and Trigger Event functions.

template<T>
void SomeTemplateFunction(T)

We also needed to create the functions for each of the TriggerEvents as those are the function pointers that will be passed
to the Abilities InputAction Gameplay Tag and let the Gameplay ability know what to do.

Kind of confusing but not too hard.

tldr: making callback functions for whether we press, release or hold down a button for an ability. we used
a template function so its all wrapped in one function and so each ability has options for each trigger event.
we need to bind the BindAbilityActions() in the player controller and remember to set the InputConfig in BP, and set the
new InputComponent as our Default inputcomponent in project settings. Our InputComponent is a child of the enhanced inputcomponent
which is the UE5 class used for enhanced inputs.

phew...