Subject: RPC & Code Clean up
We had an issue where UI popups were not firing for client players. This was due to a delegate that that got fired
on the server but didnt propagate the callback to the clients. We fixed this by making the callback function for
the delegate an RPC. We added the Client word to the name and added the _Implemenation to the Implementation of the function.
See Notion GameDev/Aura/RPC for details.

We also did some regular code clean up to shorten if/else statements for more concise code.