Created a Custom Ability Task that finds the location of the cursor in order to aim the firebolt. This works on the client but not on the server because the server is unaware of the data even though the ability is also being called on the server.
We created a Static function that returns the Ability Task type.
We override the Activate function to perform some functions when the ability is called.
We created a Dynamic Multicast Delegate that gets fires with the ability/cursor are over a valid target.
We return from the delegate an FVector of the location of the cursor & draw a debug sphere.