In the context of Unreal Engine's Gameplay Ability System (GAS), the `SendMouseCursorData()` function is used to gather and send the location of the mouse cursor. This is particularly useful in abilities that require a target location, such as a spell that is cast at the location of the mouse cursor.
The function first gets the PlayerController and uses it to get the hit result under the cursor. This hit result contains the location of the cursor in the game world. It then creates a `FGameplayAbilityTargetData_SingleTargetHit` object and sets its `HitResult` to the cursor hit result. This object is then added to a `FGameplayAbilityTargetDataHandle`.

The `FGameplayAbilityTargetDataHandle` is a container for `FGameplayAbilityTargetData` objects. It's used to pass around target data for abilities. In this case, it's used to pass the location of the mouse cursor as target data.

The `AbilitySystemComponent->ServerSetReplicatedTargetData()` function is then called with the `FGameplayAbilityTargetDataHandle` as an argument. This function is used to replicate the target data to the server. This is necessary because the server needs to know the target data to correctly execute the ability.

The `GetAbilitySpecHandle()` and `GetActivationPredictionKey()` functions are used to get the ability spec handle and the activation prediction key respectively. The ability spec handle is a unique identifier for the ability that is currently being executed. The activation prediction key is used for network prediction. It's a unique identifier for a specific activation of an ability.

Network prediction in Unreal Engine's GAS is used to reduce perceived latency. When an ability is activated, it's executed immediately on the client and the result is predicted. At the same time, the activation is sent to the server to be executed there as well. When the server has executed the ability, it sends the result back to the client. If the client's prediction was correct, there's no perceived latency. If the prediction was incorrect, the client's state is corrected.

The `FScopedPredictionWindow` object is used to scope the prediction key. This means that any replicated events that happen within the scope of the `FScopedPredictionWindow` object are associated with the prediction key. This is used to tie together the ability activation and the target data replication, so that they are considered as one atomic operation for the purpose of network prediction.